handle variable hash lengths
git version 1.8.3.1 on RHEL 7 returns the first seven characters of the
sha1.

git version 2.20.1 on Fedora 29 returns the first ten characters
instead.

As a result, when running on RHEL 7, we would always fail to find our
sha1 in the Git log, because we were looking for a 10 character string
in a set of 7 character strings.

Measure the number of characters we got back from Git, then dynamically
truncate our own sha1 hash to that length.