tar-changes: use "Update patches" message to match rdopkg
When a developer force-pushes the patches branch, rdopkg and rdopkg-tar
put slightly different versions of "Update(d) patches" in the changelog.

Update rdopkg-tar's message to match the string that rdopkg uses.

The purpose of this change is to enable other tools that read this
%changelog to match the same string across both tools.

Thanks Thomas Serlin <tserlin@redhat.com> for identifying this issue.